requirements.txt
fastapi[all]
pydantic
opencv-python==4.1.2.30
pillow==9.4.0
tensorflow==2.13.1
keras==2.13.1
segmentation-models==1.0.1
matplotlib==3.5.3